Update README.md
Here is an extended description of the code:

*Overview*

This code creates a graphical user interface (GUI) with a dropdown menu that allows users to select a file from a list of options. The GUI also includes buttons to select, open, and save the chosen file.

*Importing Libraries*

The code begins by importing the necessary libraries:

- `tkinter` (alias `tk`) for creating the GUI
- `ttk` for creating the dropdown menu

*Defining Functions*

The code defines three functions:

- `select_file()`: Retrieves the selected file from the dropdown menu and prints it to the console.
- `open_file()`: Opens the selected file and prints its contents to the console. If the file does not exist, it prints an error message.
- `save_file()`: Saves a file with the selected name and prints a success message to the console. If an error occurs, it prints an error message.

*Creating the GUI*

The code creates the GUI by:

- Creating a `tk.Tk` object to serve as the root window
- Setting the title of the root window to "File Menu"
- Creating a `tk.StringVar` object to store the selected file
- Creating a dropdown menu with the `ttk.OptionMenu` class
- Creating a frame to hold the buttons
- Creating three buttons: "Select", "Open", and "Save"

*Configuring the GUI*

The code configures the GUI by:

- Packing the dropdown menu and button frame into the root window
- Packing the buttons into the button frame
- Setting the command options for each button to call the corresponding function

*Running the GUI*

The code runs the GUI by calling the `mainloop()` method on the root window. This method starts the event loop and waits for user interactions.